directory reorganizing; including dockerfile
